Link Sparkle 1.24 via CocoaPods instead of manually
We decided to bite the bullet and move away from manually downloading,
copying, and tracking the framework to simplify our setup.

See discussion in https://github.com/Automattic/simplenote-macos/pull/803.

I considered using Swift Package Manager, which is the preferred
dependency manager for Simplenote iOS and macOS but saw two open issues
in the Sparkle repo that mentioned the need for workarounds.

- https://github.com/sparkle-project/Sparkle/issues/1689
- https://github.com/sparkle-project/Sparkle/issues/1701

It's possible that the installation would have been straightforward for
us, but I decided to stick with the tried-and-tested, one-step CocoaPods
setup instead.