Re-enable and tidy up Build Phase to set copyright year
Thanks @jkmassel; it works for me now, too.

I just updated to Xcode 12.4, but I doubt that's the reason because it
wasn't out yet at the time of the commit that disabled this. Possibly
some local messed up state in my Xcode that even removing the
DerivedData folder didn't solve. 🤷‍♂️