Update Sparkle to version 2.x commit 4767ef6 with Apple Silicon
In #803, we discussed using CocoaPods to fetch Sparkle. That's not
possible because we need support for Sandboxing and Notarization and the
version distributed via CocoaPods doesn't include it.

To support Sandboxing and Notarization, we need to use the WIP version
2.x branch. Unfortunately, the only way to use that version of Sparkle
is to fetch it and build it manually (see ./Scripts/update-sparkle.sh).

The commit I checked out was 4767ef60ece7de7404f402c23177f0810bf72418

I considered using a submodule and scripting the build process, but
decided against it because it would add complexity to the Git setup for
a relatively uncommon operation – we don't need to update Sparkle often.

For more background on Sparkle Sandboxing and Notarization, see:

- https://sparkle-project.org/documentation/sandboxing/
- https://github.com/Automattic/simplenote-macos/pull/383
- https://github.com/Automattic/simplenote-macos/pull/481