Skip possibly flaky test only on CircleCI
This way, the test runs locally.

Decided to keep it tied to CircleCI only because there is no proof that
the issue is CI related and not _Circle_CI related. This way, if we'll
ever move away from CircleCI and forget about this hack, the test will
run on the new CI.