Manually bump the version to trigger a new build
The previous CI deployment failed because CI didn't have access to the
Sentry API token. This has been addressed but we cannot simply restart
the build as a binary with that version has already been uploaded to
ASC. The reason we want a new build is to ensure we have matching dSYMs
uploaded to Sentry to track eventual crashes.